Some alterations

+++ crawl/views.py
@@ -86,44 +86,26 @@ def crawl(request):
             for u in all_links:
                 print('The links are:\n',u)
                 response = requests.get(u) #fetching each the url from the list
-                #if response.status_code == 200: #checking if the url is responsive and available but removed due to time complexity issues with heroku
-                a = urlparse(u)
-                fl=os.path.basename(a.path) #parsing the filename from the url
-                ### NLP SPACY
-                # String = "La Camera di Commercio territorialmente competente effettuerà controlli su un campione pari" #the search string from a pdf 
-                # print('The search string is :\n',String)
-                
-                # file_read=urlopen(u).read()
-                # print('The pdf with urlopen is:\n',file_read)
-                
-                # file = open(fl + ".pdf", 'wb')
-                # pdf_reader= pdf.PdfFileReader(file(u, "rb"))
-                # print('The pdf is:\n',pdf_reader)
-                
-                # NumPages = pdf_reader.getNumPages()
-                # print('The number of pages in the pdf are:\n',NumPages)
+                if response.status_code == 200: #checking if the url is responsive and available but removed due to time complexity issues with heroku
+                    a = urlparse(u)
+                    fl=os.path.basename(a.path) #parsing the filename from the url
 
-                # PageObj = pdf_reader.getPage(i)
-                # print("This is page " + str(i))
-                # Text = PageObj.extractText() #Extracting the text from the pdf page and then using the nlp to find the matching pattersns in the later stage 
-                # print(f'The extracted text from the page {i} is :\n{Text}')
-                # ResSearch = re.search(String, Text)
-                # print('The ResSearch result is :\n',ResSearch)
-                # if ResSearch =='True':
-                file_size=response.headers.get('content-length', 0)
-                content_type=response.headers.get('Content-Type', 0)
-                last_modified=response.headers.get('Last-Modified', 0)
-                expiry_date=response.headers.get('Expires', 0)
-                cache_control=response.headers.get('Cache', 0)
-                server=response.headers.get('Server', 0)
+                    file_size=response.headers.get('content-length', 0)
+                    content_type=response.headers.get('Content-Type', 0)
+                    last_modified=response.headers.get('Last-Modified', 0)
+                    expiry_date=response.headers.get('Expires', 0)
+                    cache_control=response.headers.get('Cache', 0)
+                    server=response.headers.get('Server', 0)
 
-                pub_instance = Publisher.objects.create(name=fl,links=u,file_size=file_size,content_type=content_type,last_modified_y=last_modified,expiry_date_y=expiry_date,cache_control_y=cache_control,server_y=server) #creating the entry in the database
-                pub_instance.save() #saving the info for each url to database
+                    pub_instance = Publisher.objects.create(name=fl,links=u,file_size=file_size,content_type=content_type,last_modified_y=last_modified,expiry_date_y=expiry_date,cache_control_y=cache_control,server_y=server) #creating the entry in the database
+                    pub_instance.save() #saving the info for each url to database
 
-                for row in Publisher.objects.all().reverse(): #removing all the duplicate items from the database
-                    if Publisher.objects.filter(name=row.name).count() > 1: #using name as a filter
-                        print('Found Duplicate item:\n',row.name)
-                        row.delete()
+                    for row in Publisher.objects.all().reverse(): #removing all the duplicate items from the database
+                        if Publisher.objects.filter(name=row.name).count() > 1: #using name as a filter
+                            print('Found Duplicate item:\n',row.name)
+                            row.delete()
+                else:
+                    pass
                 
             file_links=[]
             http=urllib3.PoolManager()   
@@ -200,3 +182,26 @@ def email_pdf(request):
     response = requests.request('POST', url, headers = headers, data = payload)
     return render(request, 'email_result.html')
 
+
+
+### NLP SPACY
+                # String = "La Camera di Commercio territorialmente competente effettuerà controlli su un campione pari" #the search string from a pdf 
+                # print('The search string is :\n',String)
+                
+                # file_read=urlopen(u).read()
+                # print('The pdf with urlopen is:\n',file_read)
+                
+                # file = open(fl + ".pdf", 'wb')
+                # pdf_reader= pdf.PdfFileReader(file(u, "rb"))
+                # print('The pdf is:\n',pdf_reader)
+                
+                # NumPages = pdf_reader.getNumPages()
+                # print('The number of pages in the pdf are:\n',NumPages)
+
+                # PageObj = pdf_reader.getPage(i)
+                # print("This is page " + str(i))
+                # Text = PageObj.extractText() #Extracting the text from the pdf page and then using the nlp to find the matching pattersns in the later stage 
+                # print(f'The extracted text from the page {i} is :\n{Text}')
+                # ResSearch = re.search(String, Text)
+                # print('The ResSearch result is :\n',ResSearch)
+                # if ResSearch =='True':
